+++ pkg/scheduler/plugins/factory.go
@@ -26,7 +26,7 @@ import (
 	"github.com/NVIDIA/KAI-scheduler/pkg/scheduler/plugins/gpupack"
 	"github.com/NVIDIA/KAI-scheduler/pkg/scheduler/plugins/gpusharingorder"
 	"github.com/NVIDIA/KAI-scheduler/pkg/scheduler/plugins/gpuspread"
-	"github.com/NVIDIA/KAI-scheduler/pkg/scheduler/plugins/joborder"
+	"github.com/NVIDIA/KAI-scheduler/pkg/scheduler/plugins/reflectjoborder"
 	"github.com/NVIDIA/KAI-scheduler/pkg/scheduler/plugins/kubeflow"
 	"github.com/NVIDIA/KAI-scheduler/pkg/scheduler/plugins/minruntime"
 	"github.com/NVIDIA/KAI-scheduler/pkg/scheduler/plugins/nodeavailability"
@@ -72,5 +72,5 @@ func InitDefaultPlugins() {
 	framework.RegisterPluginBuilder("snapshot", snapshot.New)
 
 	// Always register the Job Order Plugin last.
-	framework.RegisterPluginBuilder("joborder", joborder.New)
+	framework.RegisterPluginBuilder("reflectjoborder", reflectjoborder.New)
 }
